Move handle* methods to their respective locations
Fix names
